+++ docs/CHANGELOG.rst
@@ -55,6 +55,13 @@ Added
 Added
 -----
 - Support custom user model in Python processes
+- Setting ``KUBERNETES_DISPATCHER_CONFIG_LOCATION`` specifying the location of
+  the kubernetes config to load in the ``Kubernetes`` workload connector
+
+Changed
+-------
+- Retry loading ``Kubernetes`` configuration couple of times before giving up
+  in the kubernetes workload connector
 
 
 ===================